requirements.txt

Django==3.2.9
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4

